chore: address PR review feedback from tatomyr
- Inline collect-metrics.ts test helper into document-metrics.test.ts
- Use parseYaml as yaml directly instead of wrapper function
- Remove default parameter from getStylishOutput in formatter tests
- Use getMajorSpecVersion + exitWithError for spec version check
- Add explicit case 'stylish' before default in format switch
- Remove unsupported 'markdown' from score command format choices
- Add comment explaining depth=-1 initialization
- Clarify anyOf penalty and dependency terminology in docs
- Update non-oas3 rejection test for exitWithError (throws)

Made-with: Cursor
